Refactored AccommodationDetailPanel and fix failed login
