+++ pom.xml
@@ -40,13 +40,27 @@
         <dependency>
             <groupId>io.jsonwebtoken</groupId>
             <artifactId>jjwt-api</artifactId>
-            <version>0.12.6</version>
+            <version>0.13.0</version>
         </dependency>
+
         <dependency>
             <groupId>io.jsonwebtoken</groupId>
             <artifactId>jjwt-impl</artifactId>
-            <version>0.12.6</version>
+            <version>0.13.0</version>
             <scope>runtime</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>io.jsonwebtoken</groupId>
+            <artifactId>jjwt-jackson</artifactId>
+            <version>0.13.0</version>
+            <scope>runtime</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-databind</artifactId>
+            <version>2.20.0</version>
         </dependency>
 
         <!-- JUnit Dependency -->
@@ -79,12 +93,14 @@
             <version>1.4</version>
         </dependency>
 
+        <!--JX MapViewer Dependency-->
         <dependency>
             <groupId>org.jxmapviewer</groupId>
             <artifactId>jxmapviewer2</artifactId>
             <version>2.6</version>
         </dependency>
 
+        <!-- ImageIO Dependency -->
         <dependency>
             <groupId>org.sejda.imageio</groupId>
             <artifactId>webp-imageio</artifactId>
@@ -97,12 +113,14 @@
             <version>3.12.0</version>
         </dependency>
 
+        <!---Look & Feel Dependency -->
         <dependency>
             <groupId>com.formdev</groupId>
             <artifactId>flatlaf</artifactId>
             <version>3.6.1</version>
         </dependency>
 
+        <!-- JSON Dependency-->
         <dependency>
             <groupId>org.json</groupId>
             <artifactId>json</artifactId>
