+++ src/main/java/accommodationfinder/ui/AccommodationDetailPanel.java
@@ -3,24 +3,12 @@ package accommodationfinder.ui;
 import accommodationfinder.listing.Accommodation;
 import accommodationfinder.map.MapManager;
 import accommodationfinder.service.AccommodationService;
-
-import javax.imageio.IIOException;
-import javax.imageio.ImageIO;
 import javax.swing.*;
 import javax.swing.border.EmptyBorder;
-import javax.swing.event.DocumentEvent;
-import javax.swing.event.DocumentListener;
 import java.awt.*;
-import java.awt.image.BufferedImage;
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.text.NumberFormat;
 import java.time.format.DateTimeFormatter;
-import java.util.List;
 import java.util.Locale;
-import java.util.concurrent.ExecutionException;
 
 
 public class AccommodationDetailPanel extends JPanel {
@@ -29,41 +17,31 @@ public class AccommodationDetailPanel extends JPanel {
     private final MainWindow mainWindow;
     private final Long accommodationId;
     private final MapManager mapManager;
+    private Accommodation currentAccommodation;
 
 
     // UI Components
     private JLabel titleLabel;
-    private JLabel imageLabel;
     private JTextArea detailsTextArea;
-    private JButton prevImageButton;
-    private JButton nextImageButton;
-    private JLabel imageCountLabel;
-    private JTextField contactNameField;
-    private JTextField contactEmailField;
-    private JTextField contactPhoneField;
-    private JButton sendMessageButton;
-    private JLabel listerInfoLabel;
-    private JTabbedPane tabbedPane;
-    private JPanel imagePanel;
-    private JPanel locationPanel;
     private JLabel mapLabel;
+    private JPanel locationPanel;
 
-    // State
-    private List<String> currentImageUrls;
-    private int currentImageIndex = 0;
-    private Accommodation currentAccommodation;
+
+    // Sub-Panels
+    private ImageGalleryPanel imageGalleryPanel;
+    private ContactListerPanel contactListerPanel;
+
 
     // Formatters & Constants
-    private static final NumberFormat currencyFormatter = NumberFormat.getCurrencyInstance(Locale.forLanguageTag("en-ZA"));
+    private static final NumberFormat currencyFormatter = NumberFormat.getCurrencyInstance(
+            Locale.forLanguageTag("en-ZA") );
     private static final DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern("dd MMM yyyy 'at' HH:mm");
     private static final Color BACKGROUND_COLOR = new Color(253, 251, 245);
-    private static final Color PLACEHOLDER_COLOR = new Color(230, 230, 230);
-    private static final Color TEXT_COLOR = new Color(50, 50, 50);
     private static final int IMG_WIDTH = 550;
     private static final int IMG_HEIGHT = 400;
 
-    public AccommodationDetailPanel(AccommodationService accommodationService, MainWindow mainWindow,
-                                    Long accommodationId) {
+
+    public AccommodationDetailPanel(AccommodationService accommodationService, MainWindow mainWindow, Long accommodationId) {
         this.accommodationService = accommodationService;
         this.mainWindow = mainWindow;
         this.accommodationId = accommodationId;
@@ -74,118 +52,64 @@ public class AccommodationDetailPanel extends JPanel {
         setBackground(BACKGROUND_COLOR);
 
         initComponents();
-        setupContactFormListeners();
         loadAccommodationDetails();
     }
 
     private void initComponents() {
+        // Top Panel
         JPanel topPanel = new JPanel(new BorderLayout(10, 0));
         topPanel.setOpaque(false);
-
         JButton backButton = new JButton("<- Back to Main View");
-        styleButton(backButton, 14);
         backButton.addActionListener(e -> mainWindow.showMainApplicationView());
         topPanel.add(backButton, BorderLayout.WEST);
-
         titleLabel = new JLabel("Loading...", SwingConstants.LEFT);
         titleLabel.setFont(new Font("SansSerif", Font.BOLD, 24));
-        titleLabel.setBorder(new EmptyBorder(0, 10, 10, 0));
         topPanel.add(titleLabel, BorderLayout.CENTER);
-
         add(topPanel, BorderLayout.NORTH);
 
-        // Center Section
+        // Center Panel
         JPanel centerContentPanel = new JPanel(new BorderLayout(20, 10));
         centerContentPanel.setOpaque(false);
 
-        // Left Section
+        // Left Panel (Image Gallery and Details)
         JPanel leftPanel = new JPanel(new BorderLayout(10, 15));
         leftPanel.setOpaque(false);
 
-        // Create tabbed pane for Image/Location
-        tabbedPane = new JTabbedPane();
-        tabbedPane.setPreferredSize(new Dimension(IMG_WIDTH, IMG_HEIGHT + 70));
-        tabbedPane.setBackground(BACKGROUND_COLOR);
+        imageGalleryPanel = new ImageGalleryPanel(IMG_WIDTH, IMG_HEIGHT);
+        contactListerPanel = new ContactListerPanel(accommodationId);
 
-        // Image Panel (original image gallery)
-        imagePanel = new JPanel(new BorderLayout());
-        imagePanel.setBackground(PLACEHOLDER_COLOR);
-
-        imageLabel = new JLabel("Loading image...", SwingConstants.CENTER);
-        imageLabel.setOpaque(true);
-        imageLabel.setBackground(PLACEHOLDER_COLOR);
-        imageLabel.setPreferredSize(new Dimension(IMG_WIDTH, IMG_HEIGHT));
-        imageLabel.setVerticalAlignment(SwingConstants.CENTER);
-        imageLabel.setHorizontalAlignment(SwingConstants.CENTER);
-        imagePanel.add(imageLabel, BorderLayout.CENTER);
-
-        // Image Navigation Controls
-        JPanel imageNavPanel = new JPanel(new FlowLayout(FlowLayout.CENTER, 10, 5));
-        imageNavPanel.setOpaque(false);
-        prevImageButton = new JButton("< Prev");
-        styleButton(prevImageButton, 13);
-        nextImageButton = new JButton("Next >");
-        styleButton(nextImageButton, 13);
-        imageCountLabel = new JLabel("Image 0 of 0");
-        imageCountLabel.setFont(new Font("SansSerif", Font.PLAIN, 12));
-
-        prevImageButton.addActionListener(e -> showImageAtIndex(currentImageIndex - 1));
-        nextImageButton.addActionListener(e -> showImageAtIndex(currentImageIndex + 1));
-
-        imageNavPanel.add(prevImageButton);
-        imageNavPanel.add(imageCountLabel);
-        imageNavPanel.add(nextImageButton);
-        imagePanel.add(imageNavPanel, BorderLayout.SOUTH);
-
-        // Location Panel
         locationPanel = new JPanel(new BorderLayout());
         locationPanel.setBackground(BACKGROUND_COLOR);
-
-        mapLabel = new JLabel("Loading map...", SwingConstants.CENTER);
+        mapLabel = new JLabel("Select this tab to load the map", SwingConstants.CENTER);
         mapLabel.setOpaque(true);
-        mapLabel.setBackground(PLACEHOLDER_COLOR);
-        mapLabel.setPreferredSize(new Dimension(IMG_WIDTH, IMG_HEIGHT));
-        mapLabel.setVerticalAlignment(SwingConstants.CENTER);
-        mapLabel.setHorizontalAlignment(SwingConstants.CENTER);
-        mapLabel.setBorder(BorderFactory.createLineBorder(Color.LIGHT_GRAY));
+        mapLabel.setBackground(new Color(230, 230, 230));
         locationPanel.add(mapLabel, BorderLayout.CENTER);
 
-        // Add tabs
-        tabbedPane.addTab("Images", imagePanel);
+        JTabbedPane tabbedPane = new JTabbedPane();
+        tabbedPane.addTab("Images", imageGalleryPanel);
         tabbedPane.addTab("Location", locationPanel);
         leftPanel.add(tabbedPane, BorderLayout.NORTH);
 
-        // Details Area
         detailsTextArea = new JTextArea("Loading details...");
         detailsTextArea.setEditable(false);
         detailsTextArea.setLineWrap(true);
         detailsTextArea.setWrapStyleWord(true);
-        detailsTextArea.setFont(new Font("SansSerif", Font.PLAIN, 14));
-        detailsTextArea.setBorder(BorderFactory.createCompoundBorder(
-                BorderFactory.createTitledBorder("Details"),
-                new EmptyBorder(5, 5, 5, 5)
-        ));
         detailsTextArea.setBackground(BACKGROUND_COLOR);
-
+        detailsTextArea.setBorder(BorderFactory.createCompoundBorder(
+                BorderFactory.createTitledBorder("Details"), new EmptyBorder(5, 5, 5, 5)));
         JScrollPane detailsScrollPane = new JScrollPane(detailsTextArea);
-
         detailsScrollPane.setBorder(null);
         detailsScrollPane.getViewport().setBackground(BACKGROUND_COLOR);
-
         leftPanel.add(detailsScrollPane, BorderLayout.CENTER);
 
         centerContentPanel.add(leftPanel, BorderLayout.CENTER);
-
-        // Contact Form
-        JPanel contactPanel = createContactPanel();
-        centerContentPanel.add(contactPanel, BorderLayout.EAST);
+        centerContentPanel.add(contactListerPanel, BorderLayout.EAST);
 
         add(centerContentPanel, BorderLayout.CENTER);
 
-        // Add tab change listener to load map when Location tab is selected
         tabbedPane.addChangeListener(e -> {
             if (tabbedPane.getSelectedComponent() == locationPanel) {
-                if (locationPanel.getComponent(0) == mapLabel) {
+                if (locationPanel.getComponentCount() > 0 && locationPanel.getComponent(0) == mapLabel) {
                     loadInteractiveMap();
                 }
             }
@@ -203,10 +127,8 @@ public class AccommodationDetailPanel extends JPanel {
         SwingWorker<JPanel, Void> mapLoader = new SwingWorker<>() {
             @Override
             protected JPanel doInBackground() throws Exception {
-                // --- DELEGATE all map creation to the MapManager ---
                 Component mapViewer = mapManager.createMapViewer(currentAccommodation);
 
-                // --- Build the final panel to display ---
                 JPanel mapContainer = new JPanel(new BorderLayout());
                 mapContainer.add(mapViewer, BorderLayout.CENTER);
 
@@ -245,144 +167,6 @@ public class AccommodationDetailPanel extends JPanel {
         mapLoader.execute();
     }
 
-    private JPanel createContactPanel() {
-        JPanel contactPanel = new JPanel();
-        contactPanel.setLayout(new GridBagLayout());
-        contactPanel.setBorder(BorderFactory.createCompoundBorder(
-                BorderFactory.createTitledBorder("Contact Lister"),
-                new EmptyBorder(10, 10, 10, 10)
-        ));
-        contactPanel.setOpaque(false);
-        contactPanel.setBackground(BACKGROUND_COLOR);
-
-        GridBagConstraints gbc = new GridBagConstraints();
-        gbc.gridx = 0;
-        gbc.gridy = 0;
-        gbc.anchor = GridBagConstraints.WEST;
-        gbc.insets = new Insets(5, 5, 5, 5);
-
-        // Lister Info
-        listerInfoLabel = new JLabel("Listed by: Loading...");
-        listerInfoLabel.setFont(new Font("SansSerif", Font.ITALIC, 12));
-        gbc.gridwidth = 3;
-        contactPanel.add(listerInfoLabel, gbc);
-        gbc.gridwidth = 1;
-
-        gbc.gridy++;
-        contactPanel.add(new JLabel("Your Name: *"), gbc);
-        gbc.gridx = 1;
-        gbc.fill = GridBagConstraints.HORIZONTAL;
-        gbc.weightx = 1.0;
-        contactNameField = new JTextField(15);
-        contactPanel.add(contactNameField, gbc);
-        gbc.fill = GridBagConstraints.NONE;
-        gbc.weightx = 0.0;
-
-        gbc.gridy++;
-        gbc.gridx = 0;
-        contactPanel.add(new JLabel("Your Email: *"), gbc);
-        gbc.gridx = 1;
-        gbc.fill = GridBagConstraints.HORIZONTAL;
-        gbc.weightx = 1.0;
-        contactEmailField = new JTextField(15);
-        contactPanel.add(contactEmailField, gbc);
-        gbc.fill = GridBagConstraints.NONE;
-        gbc.weightx = 0.0;
-
-        gbc.gridy++;
-        gbc.gridx = 0;
-        contactPanel.add(new JLabel("Your Phone: "), gbc);
-        gbc.gridx = 1;
-        gbc.fill = GridBagConstraints.HORIZONTAL;
-        gbc.weightx = 1.0;
-        contactPhoneField = new JTextField(15);
-        contactPanel.add(contactPhoneField, gbc);
-        gbc.fill = GridBagConstraints.NONE;
-        gbc.weightx = 0.0;
-
-        gbc.gridy++;
-        gbc.gridx = 0;
-        gbc.gridwidth = 2;
-        gbc.anchor = GridBagConstraints.CENTER;
-        sendMessageButton = new JButton("Send Message");
-        styleButton(sendMessageButton, 12);
-        sendMessageButton.addActionListener(e -> handleSendMessage());
-        contactPanel.add(sendMessageButton, gbc);
-
-        // Vertical glue
-        gbc.gridy++;
-        gbc.weighty = 1.0;
-        contactPanel.add(Box.createVerticalGlue(), gbc);
-
-        return contactPanel;
-    }
-
-    // Setup listeners to enable/disable send button
-    private void setupContactFormListeners() {
-        DocumentListener listener = new DocumentListener() {
-            @Override
-            public void insertUpdate(DocumentEvent e) {
-                updateSendButtonState();
-            }
-            @Override
-            public void removeUpdate(DocumentEvent e) {
-                updateSendButtonState();
-            }
-            @Override
-            public void changedUpdate(DocumentEvent e) {
-            }
-        };
-        contactNameField.getDocument().addDocumentListener(listener);
-        contactEmailField.getDocument().addDocumentListener(listener);
-    }
-
-    // Enable send button only if Name and Email are filled
-    private void updateSendButtonState() {
-        boolean enabled = !contactNameField.getText().trim().isEmpty()
-                && !contactEmailField.getText().trim().isEmpty();
-        // email format check
-        enabled = enabled && contactEmailField.getText().trim().contains("@");
-        sendMessageButton.setEnabled(enabled);
-    }
-
-
-    private void handleSendMessage() {
-        String name = contactNameField.getText().trim();
-        String email = contactEmailField.getText().trim();
-        String phone = contactPhoneField.getText().trim();
-
-        // Re-check
-        if (name.isEmpty() || email.isEmpty()) {
-            JOptionPane.showMessageDialog(this, "Please enter your name and email.",
-
-                    "Missing Information", JOptionPane.WARNING_MESSAGE);
-            return;
-        }
-        if (!email.contains("@") || !email.contains(".")) { // Very basic validation
-            JOptionPane.showMessageDialog(this, "Please enter a valid email address.", "Invalid Email", JOptionPane.WARNING_MESSAGE);
-            return;
-        }
-
-        if (listerInfoLabel.getText().startsWith("Listed by: ") && !listerInfoLabel.getText().endsWith("Unknown User")) {
-        }
-
-        System.out.println("--- Sending Message (Placeholder) ---");
-        System.out.println("To Lister: "); // listerContactInfo
-        System.out.println("From Name: " + name);
-        System.out.println("From Email: " + email);
-        System.out.println("From Phone: " + phone);
-        System.out.println("Regarding Listing ID: " + accommodationId);
-        System.out.println("-------------------------------------");
-
-        JOptionPane.showMessageDialog(this, "Message sent to the lister" +
-                "Message to Lister", "Message Sent", JOptionPane.INFORMATION_MESSAGE);
-
-        contactNameField.setText("");
-        contactEmailField.setText("");
-        contactPhoneField.setText("");
-        updateSendButtonState();
-    }
-
     private void loadAccommodationDetails() {
         SwingWorker<Accommodation, Void> worker = new SwingWorker<>() {
             @Override
@@ -397,49 +181,21 @@ public class AccommodationDetailPanel extends JPanel {
                     if (accommodation != null) {
                         currentAccommodation = accommodation;
                         populateUI(accommodation);
-                        currentImageUrls = accommodation.getImageUrls();
-                        currentImageIndex = 0;
-                        showImageAtIndex(currentImageIndex);
+
+                        imageGalleryPanel.setImageUrls(accommodation.getImageUrls());
+                        contactListerPanel.setListerInfo(accommodation.getListedBy());
+
                     } else {
-                        displayError("Accommodation Not Found",
-                                "The requested accommodation listing could not be found.");
-                        sendMessageButton.setEnabled(false);
-                        prevImageButton.setEnabled(false);
-                        nextImageButton.setEnabled(false);
-                        imageCountLabel.setText("Image 0 of 0");
+                        displayError("Accommodation Not Found", "The requested listing could not be found.");
                     }
-                } catch (InterruptedException | ExecutionException e) {
-                    Throwable cause = e.getCause() != null ? e.getCause() : e;
-                    e.printStackTrace();
-                    displayError("Error Loading",
-                            "An error occurred while loading accommodation details:\n" + cause.getMessage());
-                    sendMessageButton.setEnabled(false);
-                    prevImageButton.setEnabled(false);
-                    nextImageButton.setEnabled(false);
-                    imageCountLabel.setText("Image 0 of 0");
                 } catch (Exception e) {
                     e.printStackTrace();
-                    displayError("UI Update Error",
-                            "An unexpected error occurred while displaying details:\n" + e.getMessage());
-                    sendMessageButton.setEnabled(false);
-                    prevImageButton.setEnabled(false);
-                    nextImageButton.setEnabled(false);
-                    imageCountLabel.setText("Image 0 of 0");
+                    displayError("Error Loading", "An error occurred while loading details: " + e.getMessage());
                 }
             }
         };
         worker.execute();
     }
-
-    private void displayError(String title, String message) {
-        titleLabel.setText(title);
-        detailsTextArea.setText(message);
-        imageLabel.setText("Error");
-        imageLabel.setIcon(null);
-        JOptionPane.showMessageDialog(AccommodationDetailPanel.this,
-                message, title, JOptionPane.ERROR_MESSAGE);
-    }
-
 
     private void populateUI(Accommodation acc) {
         titleLabel.setText(acc.getTitle());
@@ -449,7 +205,7 @@ public class AccommodationDetailPanel extends JPanel {
                 .append(acc.getPostalCode() != null ? ", " + acc.getPostalCode() : "").append("\n\n");
 
         details.append("Price: ").append(currencyFormatter.format(acc.getPrice()))
-                .append(" ").append(formatPriceFrequency(acc.getPriceFrequency())).append("\n");
+                .append(" ").append(formatPriceFrequency(acc.getPriceFrequency())).append("\n"); // You need to re-add this helper method
         details.append("Type: ").append(acc.getType()).append("\n");
         details.append("Beds: ").append(acc.getBedrooms()).append(" | Baths: ").append(acc.getBathrooms())
                 .append(" | Max Occupancy: ").append(acc.getMaxOccupancy()).append("\n\n");
@@ -475,162 +231,8 @@ public class AccommodationDetailPanel extends JPanel {
 
         detailsTextArea.setText(details.toString());
         detailsTextArea.setCaretPosition(0);
-
-        // Populate Lister Info
-        if (acc.getListedBy() != null) {
-            listerInfoLabel.setText("Listed by: " + acc.getListedBy().getFullName());
-        } else {
-            listerInfoLabel.setText("Listed by: Unknown User");
-        }
     }
 
-    // Method to show image at a specific index
-    private void showImageAtIndex(int index) {
-        if (currentImageUrls == null || currentImageUrls.isEmpty()) {
-            imageLabel.setText("No Images Available");
-            imageLabel.setIcon(null);
-            imageCountLabel.setText("Image 0 of 0");
-            prevImageButton.setEnabled(false);
-            nextImageButton.setEnabled(false);
-            return;
-        }
-
-        // Validate index
-        if (index < 0 || index >= currentImageUrls.size()) {
-            System.err.println("showImageAtIndex: Invalid index " + index);
-            return;
-        }
-
-        currentImageIndex = index;
-        loadImageAsync(currentImageIndex, IMG_WIDTH, IMG_HEIGHT);
-
-        // Update navigation state
-        int totalImages = currentImageUrls.size();
-        imageCountLabel.setText("Image " + (currentImageIndex + 1) + " of " + totalImages);
-        prevImageButton.setEnabled(currentImageIndex > 0);
-        nextImageButton.setEnabled(currentImageIndex < totalImages - 1);
-    }
-
-
-    // image loading with better scaling
-    private void loadImageAsync(int imageIndex, int targetWidth, int targetHeight) {
-        imageLabel.setIcon(null);
-        imageLabel.setText("Loading image...");
-
-        if (currentImageUrls == null || imageIndex < 0 || imageIndex >= currentImageUrls.size()) {
-            imageLabel.setText("No Image Available");
-            return;
-        }
-
-        String imageUrlString = currentImageUrls.get(imageIndex);
-        if (imageUrlString == null || imageUrlString.trim().isEmpty()) {
-            imageLabel.setText("No Image Available");
-            return;
-        }
-
-        SwingWorker<ImageIcon, Void> imageLoader = new SwingWorker<>() {
-
-            @Override
-            protected ImageIcon doInBackground() throws Exception {
-                Image scaledImage = null;
-
-                try {
-                    URL imageUrl = new URL(imageUrlString.trim());
-                    BufferedImage originalImage = null;
-                    try (InputStream is = imageUrl.openStream()) {
-                        originalImage = ImageIO.read(is);
-                    } // Stream is automatically e
-
-                    // Check if ImageIO successfully read the image including WebP via plugin
-                    if (originalImage == null) {
-                        System.err.println("Failed to load image using ImageIO (unsupported format or error): " +
-                                imageUrlString);
-                        return null;
-                    }
-
-                    // Proceed with scaling logic if image was loaded
-                    int originalWidth = originalImage.getWidth(null);
-                    int originalHeight = originalImage.getHeight(null);
-
-                    if (originalWidth <= 0 || originalHeight <= 0) {
-                        System.err.println("Invalid image dimensions for: " + imageUrlString);
-                        return null;
-                    }
-
-                    // Calculate scaled dimensions maintaining aspect ratio
-                    double scale = Math.min((double) targetWidth / originalWidth, (double) targetHeight / originalHeight);
-                    int scaledWidth = (int) (originalWidth * scale);
-                    int scaledHeight = (int) (originalHeight * scale);
-
-                    scaledWidth = Math.max(1, scaledWidth);
-                    scaledHeight = Math.max(1, scaledHeight);
-
-
-                    // Create BufferedImage for higher quality scaling
-                    BufferedImage scaledBI = new BufferedImage(scaledWidth, scaledHeight, BufferedImage.TYPE_INT_ARGB);
-                    Graphics2D g2d = scaledBI.createGraphics();
-
-                    // Apply rendering hints for better quality
-                    g2d.setRenderingHint(RenderingHints.KEY_INTERPOLATION, RenderingHints.VALUE_INTERPOLATION_BILINEAR);
-                    g2d.setRenderingHint(RenderingHints.KEY_RENDERING, RenderingHints.VALUE_RENDER_QUALITY);
-                    g2d.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
-
-                    g2d.drawImage(originalImage, 0, 0, scaledWidth, scaledHeight, null);
-                    g2d.dispose();
-
-                    return new ImageIcon(scaledBI);
-
-                } catch (MalformedURLException e) {
-                    System.err.println("Invalid image URL: " + imageUrlString + " - " + e.getMessage());
-                    return null;
-                } catch (IIOException e) { // Catch ImageIO errors
-                    System.err.println("ImageIO error loading/reading image: " + imageUrlString + " - " + e.getMessage());
-                    e.printStackTrace();
-                    return null;
-                } catch (IOException e) {
-                    System.err.println("IO error loading image stream: " + imageUrlString + " - " + e.getMessage());
-                    return null;
-                } catch (Exception e) {
-                    System.err.println("General error loading/scaling image: " + imageUrlString + " - " + e.getMessage());
-                    e.printStackTrace();
-                    return null;
-                }
-            }
-
-
-            @Override
-            protected void done() {
-                try {
-                    ImageIcon scaledIcon = get();
-                    if (scaledIcon != null) {
-                        imageLabel.setIcon(scaledIcon);
-                        imageLabel.setText(null);
-                        imageLabel.setBackground(BACKGROUND_COLOR);
-                    } else {
-                        imageLabel.setText("Image Unavailable");
-                        imageLabel.setIcon(null);
-                        imageLabel.setBackground(PLACEHOLDER_COLOR);
-                    }
-                } catch (InterruptedException | ExecutionException e) {
-                    System.err.println("Error updating detail image label: " + e.getMessage());
-                    e.printStackTrace();
-                    imageLabel.setText("Error Loading Image");
-                    imageLabel.setIcon(null);
-                    imageLabel.setBackground(PLACEHOLDER_COLOR);
-                } catch (Exception e) {
-                    System.err.println("Unexpected error updating image label: " + e.getMessage());
-                    e.printStackTrace();
-                    imageLabel.setText("Error");
-                    imageLabel.setIcon(null);
-                    imageLabel.setBackground(PLACEHOLDER_COLOR);
-                }
-            }
-        };
-        imageLoader.execute();
-    }
-
-
-    // Helper for formatting price frequency
     private String formatPriceFrequency(Accommodation.PriceFrequency frequency) {
         if (frequency == null) return "";
         return switch (frequency) {
@@ -642,15 +244,12 @@ public class AccommodationDetailPanel extends JPanel {
         };
     }
 
-    private void styleButton(JButton button, int fontSize) {
-        button.setFont(new Font("SansSerif", Font.BOLD, fontSize));
-        button.setForeground(TEXT_COLOR);
-        button.setFocusPainted(false);
+    private void displayError(String title, String message) {
+        titleLabel.setText(title);
+        detailsTextArea.setText(message);
+        JOptionPane.showMessageDialog(this, message, title, JOptionPane.ERROR_MESSAGE);
     }
 
-
-
-    // Method to return this panel for MainWindow
     public JPanel getDetailPanel() {
         return this;
     }
